% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequelae.R
\docType{data}
\name{HPV.GenitalCancerRisk}
\alias{HPV.GenitalCancerRisk}
\title{HPV.GenitalCancerRisk}
\format{
An object of class \code{list} of length 3.
}
\source{
https://en.wikipedia.org/wiki/Human_papillomavirus_infection
}
\usage{
HPV.GenitalCancerRisk
}
\description{
HPV.GenitalCancerRisk
}
\references{
DOI: 10.1016/j.vaccine.2006.05.115
}
\keyword{datasets}
