% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequelae.R
\docType{data}
\name{HPV.Sequelae}
\alias{HPV.Sequelae}
\title{HPV.Sequelae}
\format{
An object of class \code{list} of length 10.
}
\source{
https://en.wikipedia.org/wiki/Human_papillomavirus_infection
}
\usage{
HPV.Sequelae
}
\description{
HPV.Sequelae
}
\keyword{datasets}
